Day 13

+++ server/src/main/java/com/edutech/progressive/repository/WarehouseRepository.java
@@ -13,7 +13,7 @@ import org.springframework.stereotype.Repository;
 import com.edutech.progressive.entity.Warehouse;
 
 @Repository
-public interface WarehouseRepository extends JpaRepository<Warehouse, Long>{
+public interface WarehouseRepository extends JpaRepository<Warehouse, Integer>{
 
     @Query("select w from Warehouse w where w.warehouseId = :warehouseId")
     Warehouse findByWarehouseId(@Param("warehouseId") int warehouseId);
@@ -25,5 +25,8 @@ public interface WarehouseRepository extends JpaRepository<Warehouse, Long>{
     @Transactional
     @Query("delete from Warehouse w where w.supplier.supplierId = :supplierId")
     void deleteBySupplierId(@Param("supplierId")int supplierId);
+
+    
+
 
 }
